updated 'addToCart' and added 'resetPage' function
Updated 'addToCart' to hide all hat tiles when the 'add to cart' button is clicked. It also reveals the 'register items' div which is set to display: none in the html. 

Added a function to the 'cancel' button that makes the 'register items' div disappear and reveals the hat tiles again.