update javascript
I added an array called 'cartItems' to the calculateTotal function. It holds all the subtotal for each selected hat (hat.quantity * hat.price). I then use the reduce method to find the subtotal. I also updated the event listener on the 'add to cart' buttons so that the event listener is only applied to the button if that button's hat has been selected.