Update hatconfigs.js
I removed the function that was wrapping around the 'insertAdjacentHTML' method. This allowed the markup to be rendered on my page. I also added id="${hat.id} to the button. This gave each 'add to cart' button an ID attribute that I used in the event listener so that you can only use the 'add to cart' button if that particular hat has been selected.