Update checkoutscript.js
I'm now using a for loop to iterate through the inputs of each form and set the value of the billing inputs equal to the value of the shipping inputs when the checkbox is checked. If the checkbox is unchecked the billing input values are set back to an empty string and focus is placed on the first name field.