Update script.js
I added code to create a list of selected hats to the left of the 'register' area. Needs a lot of work. Currently it adds all hats each time the add to cart button is clicked. So if you choose 2 chiefs hats, click 'add to cart' button it will says Chiefs 2. If you then add 3 Brewers hats, it will say chiefs 2 chiefs 2 brewers 3. I have not added any code to the reset function to make the list disappear. I have not tried to make the list update if you remove a hat. It's also commented out right now :)